Move ProductProptype file
